Add CQL implementation
